LaunchELF v3.61 (2006.04.18)
LaunchELF v3.61 (2006.04.18)
-Added PathPad array for user defined browser shortcuts which can be
accessed in browser by pressing R2 to bring up the PathPad submenu.
Though not strictly a CNF variable, since it doesn't really configure
anything, this array is saved at the end of each CNF file, so the paths
can be reused in future sessions. (The main purpose of this is to
minimize repetitive manual browsing) NB: There is no separate save
command for this feature, so you must enter the CONFIG menu and then
exit it with 'OK', to resave a CNF with PathPad.
-Modified some menus for consistent use of Triangle button
-Modified MISC/PS2Browser to make it go directly to that browser (as
before v3.50)
-Slightly modified CD init for MISC/PS2Disc launching (eliminated a rare
error case)