LaunchELF v3.91 (2006.08.14)
LaunchELF v3.91 (2006.08.14)
-Corrected PSU name filter which was incorrect for HDD destinations
-Implemented title display mode for PSU files (using icon.sys stored in
each file)
-Implemented optional use of save title for PSU filename. This is based
on display mode, so the name chosen will be the same as that displayed.
However, since titles use Shift-JIS characters and filenames need ASCII,
the title strings will be filtered to use only normal european
characters. Any other character will be changed to underscore ('_').
-Modified JpgViewer so that rotating a picture during slideshow will
reset the timer so that a full normal viewing period is alloted after
the rotation is performed.