LaunchELF v4.42 (2010.06.12)
LaunchELF v4.42 (2010.06.12)
-Added kernel patcher and kernel patch as designed by jimmikaelkael to
make v0 japanese consoles compatible with newer models in use of
homebrews. This cures v0 problems for many other applications too
(including ESR) as the kernel remains patched for the duration of a
session (until next hard reset/power on). This patching is compatible
with that to be added into new versions of "Open PS2 Loader" too, and
includes conflict protection when multiple patchers are used. This
change has no effect at all on newer consoles, but was required to make
it possible to keep the homebrew MC modules in future versions of uLE,
and still retain uLE compatibility to all PS2 models.

LaunchELF v4.41a (2010.06.03)
-Fixed a problem with uLE initialization, causing problems with the new
MC drivers when used on a PS2 of very old model. This change means that
'initsbv_patches' is now called early in 'loadBasicModules', and nowhere
else.
-Reinstated MC drivers by jimmikaelkael, and thus FileBrowser renaming
on MCs
-Added corrections by jimmikaelkael to the EE_SIO debug output module of
ps2sdk
-added the 'sior' module to uLE with changes in "makefile",
"launchelf.h", "iopmod_name.h" and "main.c"
-Added 'SIO_DEBUG' flag to "launchelf.h" for use only in compiling
special debug versions of uLE, that use EE_SIO interface for debug
feedback instead of PS2LINK
(NB: This is required for debugging with oldest models, as PS2LINK won't
work...)
-Remerged all of the changes described for beta v4.40j described further
below