Added IOP reboot to ELF loader, for when a HDD-based target is booted.
Reboot the IOP, to leave it in a clean & consistent state.
But do not do that for boot targets on other devices, for backward-compatibility with older (homebrew) software.

It appears that it was once a thing for the booting software to leave the IOP with the required IOP modules.
This can be seen in OSDSYS v1.0x (no IOP reboot) and the mechanism to boot DVD player updates (OSDSYS will get LoadExecPS2 to load SIO2 modules).
However, it changed with the introduction of the HDD unit, as the software booted may be built with a different SDK revision.